Do not redirect to download page if no generated mockup
